+++ packages/deposit-screener/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### Unreleased
 
+- Add subgraph query functionality + test script to ensure it works
 - Add EIP712 signing logic + script for generating fixture